put in more validators for medication tab (read description)
so the only thing left in the medications tab is to put physical bounds
on what data you can input. ex, such as #characters you can put, and
what range of numbers you can enter.

so completed in the last commits was that the units now show, the fields
are OPTIONAL (as in you dont have to enter all of them). Also some
validators are put, such as you MUST put in the medication name (since
that is most important). Also, if you enter a dose/dosage/frequency then
you MUST enter a unit along with it.